--- src/auth.config.ts
@@ -1,29 +1,0 @@
-import type { NextAuthConfig } from 'next-auth';
-
-export const authConfig = {
-  pages: {
-    signIn: '/login',
-  },
-  providers: [
-    // auth.tsでCredentialsプロバイダーを追加
-  ],
-  callbacks: {
-    async redirect({ baseUrl }) {
-      return `${baseUrl}/dashboard`;
-    },
-    authorized({ auth, request: { nextUrl } }) {
-      const isLoggedIn = !!auth?.user;
-      const isOnDashboard = nextUrl.pathname.startsWith('/dashboard');
-      if (isOnDashboard) {
-        if (isLoggedIn) return true;
-        return false; // Redirect unauthenticated users to login page
-      } else if (isLoggedIn) {
-        // ログインしているユーザーがログインページにアクセスした場合、ダッシュボードにリダイレクト
-        if (nextUrl.pathname === '/login' || nextUrl.pathname === '/register') {
-          return Response.redirect(new URL('/dashboard', nextUrl));
-        }
-      }
-      return true;
-    },
-  },
-} satisfies NextAuthConfig;
